refactored mock/mock.py
