+++ autofit/mock/fixtures.py
@@ -1,7 +1,7 @@
 import autofit as af
 
 from autofit.mock.mock_model import MockClassx4
-from autofit.mock.mock import MockSamples
+from autofit.non_linear.mock.mock_samples import MockSamples
 
 
 def make_model_gaussian_x1():
