+++ test_autofit/graphical/functionality/test_visualize.py
@@ -1,6 +1,6 @@
 import autofit as af
 from autofit import graphical as g
-from autofit.mock.mock import MockAnalysis
+from autofit.non_linear.mock.mock_analysis import MockAnalysis
 
 
 class Analysis(MockAnalysis):
