+++ test_autofit/graphical/info/conftest.py
@@ -6,7 +6,7 @@ import autofit as af
 from autofit import graphical as g
 from autofit.graphical import AnalysisFactor
 from autofit.graphical.declarative.graph import GraphInfoFormatter
-from autofit.mock.mock import MockAnalysis
+from autofit.non_linear.mock.mock_analysis import MockAnalysis
 from autofit.tools.namer import namer
