+++ test_autofit/graphical/info/test_hierarchical.py
@@ -3,7 +3,7 @@ import pytest
 import autofit as af
 from autofit import graphical as g
 from autofit.graphical.declarative.graph import GraphInfoFormatter
-from autofit.mock.mock import MockAnalysis
+from autofit.non_linear.mock.mock_analysis import MockAnalysis
 
 
 @pytest.fixture(name="model_factor_1")
