+++ test_autofit/non_linear/grid/conftest.py
@@ -1,8 +1,9 @@
 import pytest
 
 import autofit as af
+
 from autofit.mock.mock_model import MockClassx2Tuple
-from autofit.mock.mock import MockSearch
+from autofit.non_linear.mock.mock_search import MockSearch
 
 @pytest.fixture(name="mapper")
 def make_mapper():
