+++ test_autofit/non_linear/samples/test_nest.py
@@ -3,7 +3,7 @@ import pytest
 import autofit as af
 
 from autofit.mock.mock_model import MockClassx4
-from autofit.mock.mock import MockNestSamples
+from autofit.non_linear.mock.mock_samples import MockNestSamples
 
 pytestmark = pytest.mark.filterwarnings("ignore::FutureWarning")
