+++ test_autofit/non_linear/samples/test_samples.py
@@ -3,7 +3,7 @@ import pytest
 
 import autofit as af
 from autofit.mock.mock_model import MockClassx2, MockClassx4
-from autofit.mock.mock import MockSamples
+from autofit.non_linear.mock.mock_samples import MockSamples
 
 pytestmark = pytest.mark.filterwarnings("ignore::FutureWarning")
