+++ test_autofit/non_linear/test_abstract_search.py
@@ -9,7 +9,7 @@ import autofit as af
 from autoconf import conf
 
 from autofit.mock.mock_model import MockClassx4, MockClassx2Tuple
-from autofit.mock.mock import MockSearch, MockSamples
+from autofit.non_linear.mock.mock_search import MockSearch, MockSamples
 
 pytestmark = pytest.mark.filterwarnings("ignore::FutureWarning")
