+++ test_autofit/non_linear/test_analysis.py
@@ -6,7 +6,7 @@ import pytest
 
 import autofit as af
 from autoconf.conf import with_config
-from autofit.mock.mock import MockSearch
+from autofit.non_linear.mock.mock_search import MockSearch
 from autofit.non_linear.analysis.multiprocessing import AnalysisPool
 from autofit.non_linear.paths.abstract import AbstractPaths
 from autofit.non_linear.paths.sub_directory_paths import SubDirectoryPaths
