+++ test_autofit/text/test_samples_text.py
@@ -4,7 +4,7 @@ import pytest
 
 import autofit as af
 
-from autofit.mock.mock import MockSamples
+from autofit.non_linear.mock.mock_samples import MockSamples
 from autofit.mock.mock_model import MockClassx2
 from autofit.mock.mock_model import MockClassx2FormatExp
 from autofit.non_linear.samples import Sample
